Update README.md
The objective of this Play Store App Review Analysis project is to deliver
insights to help understand the customer
demands better and thus helps developers
to make the right choice of the product
he/she wants to develop. We have tried to
discover the relationships among various
attributes such as which applications are
free or paid, what are the user reviews,
the user experience of the application by
rating.